+++ CyclicExhaustiveBatchQueue.cc
@@ -11,10 +11,12 @@ private:
     int ownIndex;                       // Queue's index
     cQueue buffer;                      // Buffer to save all packets
     double queueCapacity;               // Maximum space of the queue
-    double queueLen;                    // Current queue length
+    double currQueueLen;                // Current queue length
     int batchSize;                      // Batch size of this queue
     int sizeOfCurrBatch;                // Size of the current batch
     long int cycleNumber;               // Current cycle number
+    double numPacket;                   // Number of arrival packet (all packets that were generated from generator)
+    double effNumPacket;                // Efficient number of packets (that actual enter to the queue)
 
     simtime_t pollingMoment;            // The last polling moment of this queue
     simtime_t leavingMoment;            // The moment when server completes serving this queue and leaves it
@@ -23,6 +25,10 @@ private:
     cMessage *batchServDoneEvent;       // A batch of packet has been served
     std::vector<cMessage *> batch;      // Current servicing batch
 
+    // Output statistics
+    double lossRate;                    // Loss rate of the queue
+    double effArrivalRate;              // Effective arrival rate of the queue
+    double queueLen;                    // The length of queue at any time
     cStdDev queueLenAtPollingMoment;    // Queue length at its polling moment
     cStdDev waitingTime;                // The waiting time sequence of all packets
     cStdDev sojTime;                    // The sojourn time sequence of all packets
@@ -54,13 +60,19 @@ void CyclicExhaustiveBatchQueue::initialize() {
     ownIndex = par("ownIndex");
     buffer.setName("buffer");
     queueCapacity = par("queueCapacity");
-    queueLen = 0;
+    currQueueLen = 0;
     batchSize = par("batchSize");
     sizeOfCurrBatch = 0;
     cycleNumber = 0;
+    numPacket = 0;
+    effNumPacket = 0;
 
     pollingMoment = 0;
     leavingMoment = 0;
+
+    lossRate = 0;
+    effArrivalRate = 0;
+    queueLen = 0;
 
     takeBatchEvent = new cMessage("Taking a batch of packet out of the buffer");
     batchServDoneEvent = new cMessage("A batch of packet has been served");
@@ -70,6 +82,22 @@ void CyclicExhaustiveBatchQueue::initialize() {
 }
 
 void CyclicExhaustiveBatchQueue::finish() {
+    std::string lossRateNameString = "Mean loss rate in queue[" + std::to_string(ownIndex) + "]";
+    const char *lossRateName = lossRateNameString.c_str();
+    lossRate = 1 - effNumPacket/numPacket;
+    recordScalar(lossRateName, lossRate);
+
+    std::string effArrivalRateNameString = "Mean effective arrival rate in queue[" + std::to_string(ownIndex) + "]";
+    const char *effArrivalRateName = effArrivalRateNameString.c_str();
+    effArrivalRate = effNumPacket/simTime();
+    recordScalar(effArrivalRateName, effArrivalRate);
+
+    std::string queueLenNameString = "Mean queue length in queue[" + std::to_string(ownIndex) + "]";
+    const char *queueLenName = queueLenNameString.c_str();
+    // Formula Little: L_i = W_i * lambdaE_i
+    queueLen = sojTime.getMean()*effArrivalRate;
+    recordScalar(queueLenName, queueLen);
+
     std::string queueLenAtPollingMomentNameString = "Mean queue length at its polling moment in queue[" + std::to_string(ownIndex) + "]";
     const char *queueLenAtPollingMomentName = queueLenAtPollingMomentNameString.c_str();
     recordScalar(queueLenAtPollingMomentName, queueLenAtPollingMoment.getMean());
@@ -119,8 +147,8 @@ void CyclicExhaustiveBatchQueue::handleMessage(cMessage *msg) {
 //        EV << "Current cycle in Q[" << ownIndex << "] = "
 //                << cycleNumber << "\n";
 
-        queueLen = buffer.getLength();
-        if (queueLen == 0) {
+        currQueueLen = buffer.getLength();
+        if (currQueueLen == 0) {
             // Send "Queue is empty" msg to the server
             send(new cMessage("Queue is empty", ownIndex), "server$o");
             leavingMoment = simTime();
@@ -134,10 +162,10 @@ void CyclicExhaustiveBatchQueue::handleMessage(cMessage *msg) {
     // Serve the batch
     else if (msg == takeBatchEvent) {
         // Get the size of the current batch
-        if (queueLen >= batchSize)
+        if (currQueueLen >= batchSize)
             sizeOfCurrBatch = batchSize;
         else
-            sizeOfCurrBatch = queueLen;
+            sizeOfCurrBatch = currQueueLen;
 
         // Take out the batch from buffer and serve it
         for (int i = 0; i < sizeOfCurrBatch; i++) {
@@ -158,11 +186,11 @@ void CyclicExhaustiveBatchQueue::handleMessage(cMessage *msg) {
 
             // Send out the batch to the server
             send(batch[i], "server$o");
-            queueLen--;
+            currQueueLen--;
         }
 
         // Check if queue is empty
-        if (queueLen == 0) {
+        if (currQueueLen == 0) {
             send(new cMessage("Queue has been served", ownIndex), "server$o");
             leavingMoment = simTime();
         }
@@ -171,11 +199,15 @@ void CyclicExhaustiveBatchQueue::handleMessage(cMessage *msg) {
     }
 
     // Insert packet from the generator to the buffer if there is space for it
-    else
-        if (queueLen < queueCapacity) {
+    else {
+        numPacket++;
+
+        if (currQueueLen < queueCapacity) {
             buffer.insert(msg);
-            queueLen++;
+            currQueueLen++;
+            effNumPacket++;
         }
         else
             delete(msg);
+    }
 }